Add frontend option for build status output
Add a --frontend option that takes a command that will handle printing
build status.  The command will be executed with the read of a pipe on
fd 3, and each build event will cause a serialized message to be sent
on the pipe.

The frontend/ directory includes a description of the interface
between Ninja and a frontend, and an implementation of the interface
in python that mimics the native interface in ninja, which can be run
with:
ninja --frontend=frontend/native.py

(cherry-picked from commit de1875f8f868b4309befa0d671217275e2923827)

Change-Id: I473f44d95e9006282ca8829f1d613b52db01f555