Add support for --frontend_file=<file>
This will write the serialized build output into <file> (which could be
a named pipe) instead of writing it to FD3 of a subprocess.

In the Android build, a user doesn't execute ninja directly, but it's
executed by a wrapper process that interprets and prints out the build
status (from multiple ninja runs). We've been just using `--frontend=cat
<&3 >ninja_fifo` to have it written to a named pipe, but we're having
issues with that hanging on Mac. Having ninja write directly to the
named pipe seems to fix it, and is simpler overall.

Bug: 111544015
Test: switch soong_ui to use --frontend_file=
Change-Id: Id5cdc2411e635a0f8b144c0d0bfb4d2a238474b6