Add `-w missingdepfile={warn,err}`
We recently ran into a case where a tool was accidentally writing its
depfile to the wrong location. There weren't any warnings or errors when
this happened, ninja just ignored the problem, which lead to bad
incremental rebuild behaviors.

So add a warning if a command succeeds but the depfile is missing. With
`-w missingdepfile=err`, that turns into an error.

Bug: 121058584
Test: unit tests
Change-Id: I85a3ae1eb977493515d8821fc42258b0da745652