Add a thread pool that executes batches
The ninja executable defaults to creating as many threads as there are
CPUs. There is an option "-d nothreads" that disables the use of threads
(and will disable manifest splitting).

The ninja_tests executable always creates 16 threads so that a single-core
machine still exercises the thread pool.

There isn't a global thread pool. Instead, thread pools are created for a
particular task, and the worker threads are joined once the task is done.
Joining the threads reduces complications with other parts of ninja that
use fork+exec/posix_spawn/sigaction/etc.

Bug: none
Test: ninja_test
Change-Id: Icd0c982b909c649170df763950b8806a52fcba1b