Don't double-count subninja parse time
The parallel parser already avoids this problem, but it might be useful to
fix this issue in the upstream parser so the two parsers can be compared
more easily. It's less of an issue with the Android manifest tree, where
it's easy to determine the true manifest load time (halve the reported
figure), but I don't know how to determine the parse-time on projects with
extensive manifest hierarchies (e.g. Chromium).

Bug: none
Test: ninja_test
Test: see that "-d stats nothing" on AOSP reports 1 count of
  ".ninja parse" rather than 4. The run-time should be ~12 seconds, not
  ~24s.

Change-Id: I42d09cbfcbb43739ed4cb4ad207c5cb95f8c0646