Restore missing dep from depfile support
The restriction about not having phony rules that were dirty because of
a missing input was causing incremental build breaks because phony edges
are automatically created for each file listed in a depfile.

Restore the original behavior for these edges, while keeping the error
for phony rules defined in the ninja file.

Test: run ninja_tests (through build-prebuilts.sh)
Change-Id: Id41b7ecd329844a97db4d3cb9c585f26ac2774bd